rename constants to RobotConstants, add FieldConstants, move both to constants folder for organization
