+++ src/main/java/frc/robot/SubsystemFactory.java
@@ -1,8 +1,9 @@
 package frc.robot;
 
 import static edu.wpi.first.units.Units.Inches;
-import frc.robot.Constants.MotorIdConstants;
 
+import frc.robot.constants.RobotConstants;
+import frc.robot.constants.RobotConstants.MotorIdConstants;
 import frc.robot.subsystems.drive.DriveSubsystem;
 import frc.robot.subsystems.drive.SwerveModule;
 import frc.robot.subsystems.drive.SwerveModuleHardwareNEO;
@@ -103,8 +104,8 @@ public class SubsystemFactory {
                     MotorIdConstants.REAR_RIGHT_TURNING_CAN_ID,
                     REAR_RIGHT_CHASSIS_ANGULAR_OFFSET, "rear right"));
             return new DriveSubsystem(frontLeft, frontRight, rearLeft, rearRight, gyro,
-                    Constants.DriveControlConstants.BETA_XTRACK_WIDTH,
-                    Constants.DriveControlConstants.BETA_YTRACK_WIDTH);
+                    RobotConstants.DriveControlConstants.BETA_XTRACK_WIDTH,
+                    RobotConstants.DriveControlConstants.BETA_YTRACK_WIDTH);
         } else if (robotType == RobotType.MOZART) {
             frontLeft = new SwerveModule(new SwerveModuleHardwareNEO(
                     MotorIdConstants.FRONT_LEFT_DRIVING_CAN_ID,
@@ -123,8 +124,8 @@ public class SubsystemFactory {
                     MotorIdConstants.REAR_RIGHT_TURNING_CAN_ID,
                     REAR_RIGHT_CHASSIS_ANGULAR_OFFSET, "rear right"));
             return new DriveSubsystem(frontLeft, frontRight, rearLeft, rearRight, gyro,
-                    Constants.DriveControlConstants.MOZART_TRACK_WIDTH,
-                    Constants.DriveControlConstants.MOZART_TRACK_WIDTH);
+                    RobotConstants.DriveControlConstants.MOZART_TRACK_WIDTH,
+                    RobotConstants.DriveControlConstants.MOZART_TRACK_WIDTH);
         } else {
             frontLeft = new SwerveModule(new SwerveModulePlacebo());
             frontRight = new SwerveModule(new SwerveModulePlacebo());
