+++ src/main/java/frc/robot/constants/FieldConstants
@@ -1,0 +1,5 @@
+package frc.robot.constants;
+
+public final class FieldConstants {
+    
+}
