+++ src/main/java/frc/robot/subsystems/drive/DriveSubsystem.java
@@ -47,9 +47,9 @@ import edu.wpi.first.wpilibj.smartdashboard.FieldObject2d;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants;
-import frc.robot.Constants.SpeedConstants;
 import frc.robot.Robot;
+import frc.robot.constants.RobotConstants;
+import frc.robot.constants.RobotConstants.SpeedConstants;
 import frc.robot.subsystems.gyro.Gyro;
 import frc.robot.vision.VisionPoseEstimator.DriveBase;
 
@@ -225,7 +225,7 @@ public class DriveSubsystem extends SubsystemBase implements DriveBase {
                 if (Robot.isSimulation()) {
                         double angleChange = DRIVE_KINEMATICS
                                         .toChassisSpeeds(swerveModuleStates).omegaRadiansPerSecond
-                                        * (1 / Constants.SpeedConstants.MAIN_LOOP_FREQUENCY_HZ);
+                                        * (1 / RobotConstants.SpeedConstants.MAIN_LOOP_FREQUENCY_HZ);
                         lastAngle = lastAngle.plus(Rotation2d.fromRadians(angleChange));
                         gyro.setYaw(Radians.of(lastAngle.getRadians()));
                 }
